rename component's directory structure
