--- packages/loader/tsconfig.json
@@ -1,8 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "outDir": "dist",
-        "declarationDir": "dist/types"
-    },
-    "include": ["src"]
-}
