+++ stories/Ripple/loader.stories.tsx
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import notes from '../../packages/loader/README.md';
-import Loader from '../../packages/loader/src';
+import notes from '../../packages/Ripple/README.md';
+import Loader from '../../packages/Ripple/src';
 
 const colorStyle = {
   borderColor: '#FF22EF'
@@ -11,7 +11,7 @@ const sizestyle = {
   borderStyle: 'dotted'
 };
 
-storiesOf('Loader', module)
+storiesOf('Ripple', module)
   .add('with default settings', () => <Loader />, { notes })
   .add('with different color', () => <Loader {...colorStyle} />, { notes })
   .add('with different border', () => <Loader {...sizestyle} />, { notes });
